In 'Date_Calc.php', the function 'weekOfYear()' returned the ISO week, but was documented as returning the week no, counting the first full week of the year (the week containing 7th Jan.). I have implemented this functionality as:
 weekOfYear7th()

because there is no point returning the ISO week without the ISO year,
and anyway this functionality is provided by 'isoWeekDate()'.
I have also implemented:

 weekOfYear4th()
 weekOfYear1st()

which behave similary, and:

 weekOfYearAbsolute()
 weekOfMonthAbsolute()

which return the week counting the first day of the month/year as the
first week regardless of the day.
Then I have made 'weekOfYear()' an alias of 'weekOfYearAbsolute()',
because all the other week numbering definitions are confusing
and the choice of which is the 'default functionality' is a bit
controversial.

Added functions:

 daysToDayOfWeek() - for convenience
 firstDayOfYear()
 lastDayOfYear() - these two functions are designed to un-hard-code
expressions such as 'Date_Calc::dateToDays(31, 12, 2007);' because
if you wanted to inherit from this function and redefine functions
in order to modify the behaviour of the class in order to make
them specific to a particular calendar, then you might want to
use only 11 months, perhaps, or skip 10 days at some point.

Modified functions:
 isoWeekDate() - to reflect the added functions above

 julianDate()
 dayOfYear()

'julianDate()' is a badly named function because it actually returns
the day of the year, (so I have re-implemented it as 'dayOfYear()',
and in fact there is another concept called the Julian Date, which
is the number of days since 24th Nov. 4713 B.C.  I cannot find
anywhere a reference to the day of the year as being 'Julian Date'
so I assume this was a mistake.  Although 'julianDate()' now
returns the result of 'dayOfYear()' because I was worried about
breaking existing code.


git-svn-id: http://svn.php.net/repository/pear/packages/Date/trunk@245486 c90b9560-bf6c-de11-be94-00142212c4b1