Implemented Summer time (DST) in Date/TimeZone.php with data from the internet tz data-base.  It used to use get_env() and was limited to the UNIX epoch, and it used to behave a bit temperamentally on some machines.
Added the ability in PHP>5.1 to get the TZ from the local machine

Set all the private properties in 'Date' to be integers where they
are documented as being integers - the preg_match in the setDate()
function used to set them as strings, and some of the functions used
to assume they were strings (and some didn't), and I have changed
these to be consistent with each other and the documentation.

The copy() fucntion had a bug in it in PHP5 when the TZ was copied
in a shallow way (but not PHP4)

Added Date::getOrdinalSuffix()
      Date::spellNumber()
      Date::format2()
the first two being helper functions for the third, which is an
extended version of format().  The main difference is that the
format codes are variable-length, which means that it is much easier
to remember them, and as a result, more options can be implemented.
This is not intended to replace 'format()', but just to extend the
functionality of the class.
I also intended to write a function which did the reverse - converted
a date in a specified format into the object, using the same formatting
codes

Added Date::addDays(), Date::adMonths(), Date::addYears - these cannot
be specified by Date/Span.php because this works in seconds, and when
you add one month, for example, the no of seconds varies depending on what
month you are in.  Also the no of seconds is limited in the same way
as the UNIX timestamp

Added some exception throwing and handling

Added Date::secondsPastMidnight()
Added Date::getPartSecond() and Date::setPartSecond()
- these two were missing and I wanted to make the properties private

Changed the implementation of Date::setDay(), Date::setMonth(),
Date::setYear() so that it calls isValidDate() and throws an exception
if invalid

Added Date::setDayMonthYear() - necessary because of changed behaviour above

In Date/Calc.php : changed the implementation but not the behaviour of many
functions with the grand aim of reducing the definition of the
calendar to as few functions as possible.  The reason for this is that
some of the functions reflect the switch from the Julian to the Gregorian
calendar, but most of them don't, but with careful implementation, they
can be made much more consistent with each other, so that eventually
we could inherit from the class quite easily and specify which calendar
 we want by only redefining a few clearly documented functions.

I know it's not going to be useful to everyone but I thought if we
implement the calendar shift it dshould be done properly.

Also I have ironed out a few bugs I found in the Date/Calc.php implementations


git-svn-id: http://svn.php.net/repository/pear/packages/Date/trunk@245559 c90b9560-bf6c-de11-be94-00142212c4b1